changes made after video call
