+++ maps/map1.cub
@@ -8,8 +8,8 @@ F 100,100,100
 
 
 11111111111111111111111111111111111 11111  1111111
+1111E000000000000000000000000000001 10001  1000011
 11110000000000000000000000000000001 10001  1000011
-11110E00000000000000000000000000001 10001  1000011
 11111000000000000000000000000000001110001  1000011
     1000000000000000000000000000000000001  10000111111
     1000000000000000000000000000000000001  10000000011
