+++ srcs/fps.c
@@ -42,3 +42,13 @@ char	*set_fps_string(char *str1, t_cub *cub, int mode)
 		return (NULL);
 	return (new);
 }
+
+unsigned long long	get_time_micros(void)
+{
+	struct timeval		tv;
+	unsigned long long	curtime;
+
+	gettimeofday(&tv, NULL);
+	curtime = (unsigned long long)((tv.tv_sec * 1000000 + tv.tv_usec));
+	return (curtime);
+}
