+++ srcs/hooks/key_hook.c
@@ -1,0 +1,29 @@
+#include "cub.h"
+
+int	key_hook(int keycode, t_cub *cub)
+{
+	double	edge;
+	double	move_speed;
+	double	rot_speed;
+
+	edge = 1.1;
+	move_speed = 0.5;
+	rot_speed = 0.08;
+	if (keycode == ESC_KEY)
+		graceful_exit(cub);
+	else if (keycode == W_KEY || keycode == UP_KEY)
+		move_forward(cub, edge, move_speed);
+	else if (keycode == S_KEY || keycode == DOWN_KEY)
+		move_backward(cub, edge, move_speed);
+	else if (keycode == LEFT_KEY)
+		rotate_left(cub, rot_speed);
+	else if (keycode == RIGHT_KEY)
+		rotate_right(cub, rot_speed);
+	else if (keycode == A_KEY)
+		move_left(cub, edge, move_speed);
+	else if (keycode == D_KEY)
+		move_right(cub, edge, move_speed);
+	raycast(cub);
+	mlx_put_image_to_window(cub->mlx, cub->window, cub->img, 0, 0);
+	return (0);
+}
