--- srcs/time_utils.c
@@ -1,11 +1,0 @@
-#include "cub.h"
-
-unsigned long long	get_time_micros(void)
-{
-	struct timeval		tv;
-	unsigned long long	curtime;
-
-	gettimeofday(&tv, NULL);
-	curtime = (unsigned long long)((tv.tv_sec * 1000000 + tv.tv_usec));
-	return (curtime);
-}
